% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW3.R
\name{linear_regression}
\alias{linear_regression}
\title{linear_regression function}
\usage{
linear_regression(data, depend_var, independ_vars)
}
\arguments{
\item{data}{dataset. This input value is any dataset including several variables.}

\item{depend_var}{dependent variable. A variable would like to measure,
which depends on the independent variable.}

\item{independ_vars}{independent variable. A variable in a model used to forecast
the outcome of dependent variable based on its value.}
}
\value{
a list of relative statistical values of linear regression through calculation.
}
\description{
This function helps to calculate the relative statistical values of linear regression.
}
\examples{
linear_regression(mtcars, "mpg", c("wt", "hp", "qsec"))

}
